+++ app/views/panel.js
