+++ app/views/ui/timeline-layer.js
