+++ styles/stylesheet.css
@@ -64,7 +64,10 @@ article, aside, details, figcaption, figure, footer, header, hgroup, menu, nav, 
   display: block;
 }
 
-/* line 14, stylesheet.sass */
+/* ==============================
+ * ===- BASE
+ * ============================== */
+/* line 18, stylesheet.sass */
 * {
   -webkit-box-sizing: border-box;
   -moz-box-sizing: border-box;
@@ -78,12 +81,12 @@ article, aside, details, figcaption, figure, footer, header, hgroup, menu, nav, 
   user-select: none;
 }
 
-/* line 26, stylesheet.sass */
+/* line 30, stylesheet.sass */
 html {
   height: 100%;
 }
 
-/* line 29, stylesheet.sass */
+/* line 33, stylesheet.sass */
 body {
   height: 100%;
   background: #1d1d1d;
@@ -92,7 +95,7 @@ body {
   font-weight: 300;
 }
 
-/* line 36, stylesheet.sass */
+/* line 40, stylesheet.sass */
 header {
   position: absolute;
   top: 0;
@@ -101,7 +104,7 @@ header {
   min-height: 50px;
 }
 
-/* line 43, stylesheet.sass */
+/* line 47, stylesheet.sass */
 h1 {
   color: #00ccff;
   font-size: 30px;
@@ -110,13 +113,13 @@ h1 {
   text-indent: 40px;
 }
 
-/* line 50, stylesheet.sass */
+/* line 54, stylesheet.sass */
 h2 {
   font-size: 30px;
   margin: 10px 0 10px 20px;
 }
 
-/* line 54, stylesheet.sass */
+/* line 58, stylesheet.sass */
 .video {
   position: absolute;
   top: 50px;
@@ -124,6 +127,9 @@ h2 {
   height: 80%;
 }
 
+/* ==============================
+ * ===- IMPORT
+ * ============================== */
 /* line 1, ui/footer.sass */
 footer {
   background: #333333;
@@ -255,22 +261,22 @@ footer .bar.effects .layer .part {
   filter: progid:DXImageTransform.Microsoft.gradient(startColorstr='#4cffa7', endColorstr='#36d486',GradientType=0 );
 }
 
-/* line 10, panels/panel.sass */
+/* line 10, ui/panel.sass */
 .ui-panel {
   width: 280px;
   position: absolute;
   top: 0;
   bottom: 50px;
 }
-/* line 16, panels/panel.sass */
+/* line 16, ui/panel.sass */
 .ui-panel.left {
   left: 20px;
 }
-/* line 19, panels/panel.sass */
+/* line 19, ui/panel.sass */
 .ui-panel.right {
   right: 20px;
 }
-/* line 22, panels/panel.sass */
+/* line 22, ui/panel.sass */
 .ui-panel ul {
   width: 100%;
   position: absolute;
@@ -279,54 +285,53 @@ footer .bar.effects .layer .part {
   overflow: hidden;
   overflow-y: auto;
 }
-/* line 31, panels/panel.sass */
+/* line 31, ui/panel.sass */
 .ui-panel ul li {
   position: relative;
   width: 100%;
   height: 155px;
   margin: 10px 0 0;
 }
-/* line 37, panels/panel.sass */
+/* line 37, ui/panel.sass */
 .ui-panel ul li:first-child {
   margin: 0;
 }
 
-/* line 1, panels/medialibrary.sass */
-.file-selector {
+/* line 3, ui/library.sass */
+#library .file-selector {
   background: #333333 url("../../images/whitecross.png") no-repeat center center;
   width: 280px;
 }
-/* line 5, panels/medialibrary.sass */
-.file-selector form, .file-selector input {
+/* line 7, ui/library.sass */
+#library .file-selector form, #library .file-selector input {
   cursor: pointer;
   opacity: 0;
   width: 280px;
   height: 100%;
 }
-/* line 11, panels/medialibrary.sass */
-.file-selector:hover {
+/* line 13, ui/library.sass */
+#library .file-selector:hover {
   background-color: #00ccff;
 }
-
-/* line 14, panels/medialibrary.sass */
-.mediaLibraryItem {
+/* line 16, ui/library.sass */
+#library .library-item {
   width: 280px;
   height: 155px;
 }
-/* line 18, panels/medialibrary.sass */
-.mediaLibraryItem > * {
+/* line 20, ui/library.sass */
+#library .library-item > * {
   position: absolute;
 }
-/* line 21, panels/medialibrary.sass */
-.mediaLibraryItem h3 {
+/* line 23, ui/library.sass */
+#library .library-item h3 {
   opacity: 0;
   background: rgba(0, 0, 0, 0.5);
   width: 100%;
   line-height: 1.5em;
   text-indent: 5px;
 }
-/* line 28, panels/medialibrary.sass */
-.mediaLibraryItem button {
+/* line 30, ui/library.sass */
+#library .library-item button {
   background: #00ccff;
   color: black;
   border-radius: 3px;
@@ -336,24 +341,24 @@ footer .bar.effects .layer .part {
   margin: 125px 5px 0 -2px;
   position: relative;
 }
-/* line 38, panels/medialibrary.sass */
-.mediaLibraryItem img, .mediaLibraryItem video {
+/* line 40, ui/library.sass */
+#library .library-item img, #library .library-item video {
   width: 100%;
   height: auto;
 }
-/* line 45, panels/medialibrary.sass */
-.mediaLibraryItem:hover h3 {
+/* line 47, ui/library.sass */
+#library .library-item:hover h3 {
   opacity: 1;
 }
 
-/* line 64, stylesheet.sass */
+/* line 72, stylesheet.sass */
 #composition {
   position: absolute;
   left: 310px;
   right: 310px;
   height: 80%;
 }
-/* line 71, stylesheet.sass */
+/* line 79, stylesheet.sass */
 #composition .controls {
   z-index: 1;
   position: absolute;
@@ -362,13 +367,13 @@ footer .bar.effects .layer .part {
   left: 20px;
 }
 
-/* line 78, stylesheet.sass */
+/* line 86, stylesheet.sass */
 #player {
   width: 100%;
   pointer-events: none;
 }
 
-/* line 83, stylesheet.sass */
+/* line 91, stylesheet.sass */
 button {
   width: 40px;
   height: 40px;
@@ -376,18 +381,14 @@ button {
   background: center center no-repeat;
   cursor: pointer;
 }
-/* line 89, stylesheet.sass */
+/* line 97, stylesheet.sass */
 button:hover {
   opacity: 0.8;
 }
 
-/* line 92, stylesheet.sass */
+/* line 100, stylesheet.sass */
 #fps {
   position: absolute;
   top: 10px;
   right: 10px;
 }
-
-/* ==============================
- * ===- PANEL - LIBRARY
- * ============================== */
