+++ styles/stylesheet.sass
@@ -11,6 +11,10 @@ $panelWidth: 280px
 
 $titleBlue: #0CF
 
+/* ==============================
+ * ===- BASE
+ * ==============================
+
 *
   -webkit-box-sizing: border-box
   -moz-box-sizing: border-box
@@ -20,7 +24,7 @@ $titleBlue: #0CF
   -webkit-user-select: none
   -khtml-user-select: none
   -moz-user-select: moz-none
-  -ms-user-select: none
+  -ms-user-select: none 
   user-select: none
 
 html
@@ -56,10 +60,14 @@ h2
   top: $headerHeight
   width: 100%
   height: $panelHeightPct
+  
+/* ==============================
+ * ===- IMPORT
+ * ==============================
 
 @import ui/footer
-@import "panels/panel"
-@import "panels/medialibrary"
+@import ui/panel
+@import ui/library
 
 #composition
   position: absolute
@@ -93,10 +101,3 @@ button
   position: absolute
   top: 10px
   right: 10px
-
-
-/* ==============================
- * ===- PANEL - LIBRARY
- * ==============================
-
-//#library.ui-panel li.file-selector input
