+++ styles/ui/panel.css
