+++ styles/ui/panel.sass
